requirements.txt
Flask
python-dotenv
openai
langchain
langgraph
langchain-openai
chromadb
psycopg2-binary 
langchain-community
pandas
pyodbc